functions/requirements.txt
# Function dependencies, for example:
# package>=version

beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.16.0
MarkupSafe==1.1.1
nltk==3.5
openai==0.2.4
readability-lxml==0.8.1
regex==2020.7.14
requests==2.24.0
soupsieve==2.0.1
tqdm==4.48.2
urllib3==1.25.10
Werkzeug==1.0.1
